[AAP-25503] Refactor instance tests
Update instance tests and instance groups.
